requirements.txt
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
coverage==6.3.2
crispy-bootstrap5==0.6
Django==4.0.3
django-coverage-plugin==2.0.2
django-crispy-forms==1.14.0
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.3.4
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.7
pytest==7.1.1
pytest-django==4.5.2
pytest-sugar==0.9.4
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
tomli==2.0.1
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
